Refactor TopicProducerActor: use backoffSource in the actor
