+++ service/core/kafka/client/src/main/resources/reference.conf
@@ -18,6 +18,9 @@ lagom.broker.kafka {
 
   client {
     default {
+      # how long should we wait when retrieving the last known offset
+      offset-timeout = 5s
+
       # Exponential backoff for failures
       failure-exponential-backoff {
         # minimum (initial) duration until processor is started again
